order created from the cart list of the user, cleared cart list. Ref #34

+++ backend/src/main/java/com/cloud/webshop/service/impl/CartServiceImpl.java
@@ -59,7 +59,7 @@ public class CartServiceImpl implements CartService {
         CartItemResponse cartItemResponse = CartItemResponse.toCartItemResponse(product, quantity);
         cartItemResponse.setProduct(productResponse);
         cartItemResponse.setQuantity(quantity);
-        cartItemResponse.setTotalPrice(product.getPrice().doubleValue() * quantity);
+        cartItemResponse.setTotalPrice(product.getPrice() * quantity);
 
         return cartItemResponse;
     }
